requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
scikit-learn
pillow
sqlalchemy
databases
open_clip_torch
